requirements.txt
aws-cdk-lib==2.48.0
constructs>=10.0.0,<11.0.0
boto3==1.26.143
pyyaml==6.0
